+++ src/control_plane/migration/redis/mod.rs
@@ -3,6 +3,10 @@ mod tests;
 
 mod backup_redis_prefix;
 mod redis_backup_options;
+mod redis_restore_options;
+mod restore_redis_prefix;
 
 pub(crate) use backup_redis_prefix::backup_redis_prefix;
 pub(crate) use redis_backup_options::RedisBackupOptions;
+pub(crate) use redis_restore_options::RedisRestoreOptions;
+pub(crate) use restore_redis_prefix::restore_redis_prefix;